requirements.txt
kivy
cython
pygame
pillow